feat(mcp): add patch_dashboard, get_dashboard_tile, search_dashboards tools (HDX-4139)
Add three new MCP dashboard tools for more granular operations:

- hyperdx_get_dashboard_tile: retrieve a single tile by tileId without
  loading the full dashboard
- hyperdx_patch_dashboard: update dashboard name/tags and/or replace a
  single tile by tileId in one call, preserving all other tiles and
  falling back to existing layout when omitted
- hyperdx_search_dashboards: search dashboards by name (case-insensitive)
  and/or tags

Additional improvements:
- Fix empty parameter schema on patch_dashboard and search_dashboards
  caused by Zod .refine() wrapping (moved cross-field validation to
  handler body so inputSchema stays a plain z.object)
- Add 'prefer patch_dashboard' hint to save_dashboard description
- Document Lucene substring matching limitation prominently in tool
  descriptions and query guide prompt (field:value is ilike not equality,
  field:val* is prefix-within-substring not true prefix)
- Split monolithic dashboards.test.ts (3040 lines) into 8 focused test
  files under __tests__/dashboards/
