+++ packages/api/jest.config.js
@@ -10,7 +10,7 @@ module.exports = {
   testEnvironment: 'node',
   verbose: true,
   rootDir: './src',
-  testMatch: ['**/__tests__/*.test.ts?(x)'],
+  testMatch: ['**/__tests__/**/*.test.ts?(x)'],
   testTimeout: 30000,
   moduleNameMapper: {
     '@/(.*)$': '<rootDir>/$1',
