+++ packages/api/src/mcp/__tests__/dashboards/queryTile.test.ts
@@ -1,0 +1,98 @@
+import { callTool, getFirstText } from '../mcpTestUtils';
+import { setupDashboardTests } from './setup';
+
+describe('MCP Dashboard Tools - hyperdx_query_tile', () => {
+  const ctx = setupDashboardTests();
+
+  it('should return error for non-existent dashboard', async () => {
+    const result = await callTool(ctx.client!, 'hyperdx_query_tile', {
+      dashboardId: '000000000000000000000000',
+      tileId: 'some-tile-id',
+    });
+
+    expect(result.isError).toBe(true);
+    expect(getFirstText(result)).toContain('not found');
+  });
+
+  it('should return error for non-existent tile', async () => {
+    const sourceId = ctx.traceSource._id.toString();
+    const createResult = await callTool(ctx.client!, 'hyperdx_save_dashboard', {
+      name: 'Tile Query Test',
+      tiles: [
+        {
+          name: 'My Tile',
+          config: {
+            displayType: 'number',
+            sourceId,
+            select: [{ aggFn: 'count' }],
+          },
+        },
+      ],
+    });
+    const dashboard = JSON.parse(getFirstText(createResult));
+
+    const result = await callTool(ctx.client!, 'hyperdx_query_tile', {
+      dashboardId: dashboard.id,
+      tileId: 'non-existent-tile-id',
+    });
+
+    expect(result.isError).toBe(true);
+    expect(getFirstText(result)).toContain('Tile not found');
+  });
+
+  it('should return error for invalid time range', async () => {
+    const sourceId = ctx.traceSource._id.toString();
+    const createResult = await callTool(ctx.client!, 'hyperdx_save_dashboard', {
+      name: 'Time Range Test',
+      tiles: [
+        {
+          name: 'Tile',
+          config: {
+            displayType: 'number',
+            sourceId,
+            select: [{ aggFn: 'count' }],
+          },
+        },
+      ],
+    });
+    const dashboard = JSON.parse(getFirstText(createResult));
+
+    const result = await callTool(ctx.client!, 'hyperdx_query_tile', {
+      dashboardId: dashboard.id,
+      tileId: dashboard.tiles[0].id,
+      startTime: 'not-a-date',
+    });
+
+    expect(result.isError).toBe(true);
+    expect(getFirstText(result)).toContain('Invalid');
+  });
+
+  it('should execute query for a valid tile', async () => {
+    const sourceId = ctx.traceSource._id.toString();
+    const createResult = await callTool(ctx.client!, 'hyperdx_save_dashboard', {
+      name: 'Query Tile Test',
+      tiles: [
+        {
+          name: 'Count Tile',
+          config: {
+            displayType: 'number',
+            sourceId,
+            select: [{ aggFn: 'count' }],
+          },
+        },
+      ],
+    });
+    const dashboard = JSON.parse(getFirstText(createResult));
+
+    const result = await callTool(ctx.client!, 'hyperdx_query_tile', {
+      dashboardId: dashboard.id,
+      tileId: dashboard.tiles[0].id,
+      startTime: new Date(Date.now() - 24 * 60 * 60 * 1000).toISOString(),
+      endTime: new Date().toISOString(),
+    });
+
+    // Should succeed (may have empty results since no data inserted)
+    expect(result.isError).toBeFalsy();
+    expect(result.content).toHaveLength(1);
+  });
+});
