+++ packages/api/src/mcp/tools/dashboards/index.ts
@@ -3,8 +3,11 @@ import type { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 import type { McpContext, ToolDefinition } from '../types';
 import { registerDeleteDashboard } from './deleteDashboard';
 import { registerGetDashboard } from './getDashboard';
+import { registerGetDashboardTile } from './getDashboardTile';
+import { registerPatchDashboard } from './patchDashboard';
 import { registerQueryTile } from './queryTile';
 import { registerSaveDashboard } from './saveDashboard';
+import { registerSearchDashboards } from './searchDashboards';
 
 export * from './schemas';
 
@@ -13,8 +16,11 @@ const dashboardsTools: ToolDefinition = (
   context: McpContext,
 ) => {
   registerGetDashboard(server, context);
+  registerGetDashboardTile(server, context);
   registerSaveDashboard(server, context);
+  registerPatchDashboard(server, context);
   registerDeleteDashboard(server, context);
+  registerSearchDashboards(server, context);
   registerQueryTile(server, context);
 };
